backend/requirements.txt
asgiref==3.10.0
astroid==3.3.11
colorama==0.4.6
contourpy==1.3.3
cycler==0.12.1
dill==0.4.0
Django==5.2.7
django-bower==5.2.0
django-cors-headers==4.9.0
django-mathfilters==1.0.0
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
et_xmlfile==2.0.0
fonttools==4.60.1
isort==6.1.0
kiwisolver==1.4.9
matplotlib==3.10.7
mccabe==0.7.0
numpy==2.3.4
openpyxl==3.1.5
packaging==25.0
pandas==2.3.3
Pillow==12.0.0
<<<<<<< HEAD
=======
pip==25.3
>>>>>>> tilines_backend
platformdirs==4.5.0
psycopg2-binary==2.9.11
PyJWT==2.10.1
pylint==3.3.9
pylint-django==2.6.1
pylint-plugin-utils==0.9.0
pyparsing==3.2.5
python-dateutil==2.9.0.post0
pytz==2025.2
setuptools==80.9.0
six==1.17.0
sqlparse==0.5.3
tomlkit==0.13.3
tzdata==2025.2
xlrd==1.2.0
xlwt==1.3.0
